+++ Dockerfile.graphql
@@ -3,7 +3,7 @@ LABEL maintainer="Nyaomaru<nyaonyao0725@gmail.com>"
 
 WORKDIR /app
 
-RUN git clone https://github.com/nyaomaru/graphql_sample . && \
+RUN git clone https://github.com/nyaomaru/graphql-sample . && \
     yarn
 
 EXPOSE 4000
