+++ Dockerfile.swagger
@@ -3,7 +3,7 @@ LABEL maintainer="Nyaomaru<nyaonyao0725@gmail.com>"
 
 WORKDIR /app
 
-RUN git clone https://github.com/nyaomaru/react_vite_swagger . && \
+RUN git clone https://github.com/nyaomaru/react-vite-swagger . && \
     npm install
 
 EXPOSE 8080
