+++ src/routes/customer/list/-components/CustomerList.tsx
@@ -5,7 +5,7 @@ import { ErrorAlert } from '@/components/atoms/ErrorAlert';
 import { SimpleButton } from '@/components/atoms/SimpleButton';
 import { CustomerQueries } from '@/queries/customer.queries';
 
-import { PATH } from '@/constant/routes';
+import { PATH } from '@/constants/routes';
 
 import { ButtonStyle } from '@/routes/customer/-components/Customer.css';
 import { CustomerTable } from '@/routes/customer/-components/CustomerTable';
